fix build on 32-bit systems
Addresses https://github.com/facebook/libphenom/issues/13
The underlying issue is that 32-bit systems don't have 64-bit CAS
primitives so we need to size our ring bits to the system pointer
size instead.

I did some light testing of this by amending tpool.t to use 64 threads
and it seemed to work (the concern would be whether we overflow
MAX_RINGS)

+++ include/phenom/counter.h
@@ -381,7 +381,7 @@ uint8_t ph_counter_scope_get_num_slots(
 /* an iterator for scopes */
 struct ph_counter_scope_iterator {
   void *ptr;
-  uint64_t offset;
+  intptr_t offset;
 };
 typedef struct ph_counter_scope_iterator ph_counter_scope_iterator_t;
